[Cart] Extract common summary action
