+++ src/Resources/config/services/actions/cart.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+    <services>
+        <service id="sylius.shop_api_plugin.controller.cart.summarize_action"
+                 class="Sylius\ShopApiPlugin\Controller\Cart\SummarizeAction"
+        >
+            <argument type="service" id="sylius.repository.order" />
+            <argument type="service" id="fos_rest.view_handler" />
+            <argument type="service" id="sylius.shop_api_plugin.factory.cart_view_factory" />
+        </service>
+    </services>
+</container>
