+++ src/Resources/config/services/actions/checkout.xml
@@ -2,13 +2,6 @@
 
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
-        <service id="sylius.shop_api_plugin.controller.checkout.summarize_action"
-                 class="Sylius\ShopApiPlugin\Controller\Checkout\SummarizeAction"
-        >
-            <argument type="service" id="sylius.repository.order" />
-            <argument type="service" id="fos_rest.view_handler" />
-            <argument type="service" id="sylius.shop_api_plugin.factory.cart_view_factory" />
-        </service>
         <service id="sylius.shop_api_plugin.controller.checkout.address_action"
                  class="Sylius\ShopApiPlugin\Controller\Checkout\AddressAction"
         >
