+++ src/Resources/config/services/controllers.xml
@@ -3,6 +3,7 @@
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <imports>
         <import resource="actions/checkout.xml"/>
+        <import resource="actions/cart.xml"/>
     </imports>
     <services>
         <service id="sylius.shop_api_plugin.controller.taxon_controller"
